requirements.txt
numpy
xgboost
matplotlib
scikit-learn
perlin_noise
Pillow